Move the DisplayFieldNameAttribute definition into a separate file
